requirements.txt
streamlit
opencv-python
ultralytics
numpy
pytesseract
pillow